% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{excludeDecision}
\alias{excludeDecision}
\title{Exclude a decision from a task}
\usage{
excludeDecision(decision, connStr)
}
\arguments{
\item{decision}{A decision id.}

\item{connStr}{A connection string.}
}
\value{
updated count
}
\description{
Exclude a decision from a task
}
