% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{excludeDecisions}
\alias{excludeDecisions}
\title{Exclude decisions from a task}
\usage{
excludeDecisions(decisions, connStr)
}
\arguments{
\item{decisions}{A vector of decision ids.}

\item{connStr}{A connection string.}
}
\value{
updated count
}
\description{
Exclude decisions from a task
}
