% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{excludeJudge}
\alias{excludeJudge}
\title{Exclude a judge from a task}
\usage{
excludeJudge(judge_id, modTask = FALSE, connStr)
}
\arguments{
\item{judge_id}{The judge id}

\item{modTask}{Is this a mod task, TRUE or FALSE?}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Exclude a judge from a task
}
\examples{
excludeJudge('o544AXuwEBADM9Lrk', FALSE,'mongodb://')
}
