% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judges.R
\name{excludeJudgeMod}
\alias{excludeJudgeMod}
\title{Exclude a judge from a moderation task}
\usage{
excludeJudgeMod(judge_id, connStr)
}
\arguments{
\item{judge_id}{The judge id}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Exclude a judge from a moderation task
}
\examples{
excludeJudge('o544AXuwEBADM9Lrk', 'mongodb://')
}
