% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{getAgeToScaledScore}
\alias{getAgeToScaledScore}
\title{Get age to scaled score look up table}
\usage{
getAgeToScaledScore(writingAgeSet, connStr)
}
\arguments{
\item{writingAgeSet}{The writing age set}

\item{connStr}{A connection string.}
}
\value{
A data frame with daysAtTest, scaledScore, yearsAtTest and writingAgeSet
}
\description{
Get the age.to.scaled.score look up table for a given writing age set
}
\examples{
\dontrun{
getAgeToScaledScore('APW 2024-2025 Standardisation', 'mongodb://localhost')
}
}
