% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judges.R
\name{getAudioTranscripts}
\alias{getAudioTranscripts}
\title{Get audio transcripts}
\usage{
getAudioTranscripts(tasks, connStr)
}
\arguments{
\item{tasks}{A vector of tasks}

\item{connStr}{A connection string.}
}
\value{
A data frame with audio transcripts
}
\description{
Get audio transcripts
}
\examples{
getAudioTranscripts(tasks,connStr)
}
