% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getDecisions}
\alias{getDecisions}
\title{Get decisions from a task or tasks}
\usage{
getDecisions(taskName, connStr)
}
\arguments{
\item{taskName}{The task name, tasks will be matched if their name includes the taskName string.}

\item{connStr}{A connection string.}
}
\value{
A data frame with decisions.
}
\description{
Get decisions from a task or tasks
}
\examples{
getDecisions('Sharing Standards 2017-2018 Year 3 9365206', 'mongodb://')
}
