% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{getEditor}
\alias{getEditor}
\title{Get the contents of the editor}
\usage{
getEditor(task, connStr)
}
\arguments{
\item{task}{Task Id}
}
\value{
A data frame with the contents of the editor
}
\description{
Get the contents of the editor
}
\examples{
getEditor(localTask,connStr)
}
