% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getImages}
\alias{getImages}
\title{Get image paths}
\usage{
getImages(task, connStr)
}
\arguments{
\item{task}{The id of the local task}

\item{connStr}{A connection string}
}
\value{
A data frame with details from the scans.Pages collection
}
\description{
Get image paths
}
\examples{
getImages('abc', 'mongodb://') 
}
