% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{getInvoiceByNumber}
\alias{getInvoiceByNumber}
\title{Get invoice details from an invoice number}
\usage{
getInvoiceByNumber(invoice, connStr)
}
\arguments{
\item{invoice}{The invoice number}

\item{connStr}{A connection string.}
}
\value{
A data frame with invoice
}
\description{
Get invoice details from an invoice number
}
\examples{
getInvoiceByNumber(123, 'mongodb://')
}
