% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{getInvoicesByDfe}
\alias{getInvoicesByDfe}
\title{Get invoice details from dfe number}
\usage{
getInvoicesByDfe(dfe, connStr)
}
\arguments{
\item{dfe}{The dfe number}

\item{connStr}{A connection string.}
}
\value{
A data frame with invoices
}
\description{
Get invoice details from dfe number
}
\examples{
getInvoicesByDfe(123, 'mongodb://')
}
