% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judges.R
\name{getJudgeComments}
\alias{getJudgeComments}
\title{Get judge comments}
\usage{
getJudgeComments(tasks, connStr)
}
\arguments{
\item{tasks}{A vector of tasks}

\item{connStr}{A connection string.}
}
\value{
A data frame with comments
}
\description{
Get judge comments
}
\examples{
getJudgeComments(tasks,connStr)
}
