% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getJudges}
\alias{getJudges}
\title{Get judge details}
\usage{
getJudges(taskId, connStr, infit = 0, comparisons = 10000)
}
\arguments{
\item{taskId}{Task Id}

\item{connStr}{A connection string.}

\item{infit}{Only judges equal to or above this infit}

\item{comparisons}{Only judges with equal of fewer than these comparisons}
}
\value{
A data frame with user details.
}
\description{
Get judge details
}
\examples{
getJudges('ocjZYZAYLuZJyAv4h', 'mongodb://')
}
