% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judges.R
\name{getJudgesBySyllabus}
\alias{getJudgesBySyllabus}
\title{Get judges for a syllabus}
\usage{
getJudgesBySyllabus(syllabus, connStr)
}
\arguments{
\item{syllabus}{The syllabus id}

\item{connStr}{A connection string.}
}
\value{
A data frame with judge details.
}
\description{
Get judges for a syllabus
}
\examples{
getJudgesBySyllabus(syllabusid, 'mongodb://')
}
