% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getJudging}
\alias{getJudging}
\title{Get judging pair for a judge id}
\usage{
getJudging(judge, connStr)
}
\arguments{
\item{connStr}{A connection string.}

\item{id}{The judge id.}
}
\value{
A named list with the id of the pair & name of the left and right script
}
\description{
Get judging pair for a judge id
}
\examples{
getJudging('efY4MWWcNuoHYPZ5B', 'mongodb://')
}
