% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getLevels}
\alias{getLevels}
\title{Get level names and cut off values for a task}
\usage{
getLevels(taskName, connStr)
}
\arguments{
\item{taskName}{The task name. Doesn't support regular expressions.}

\item{connStr}{A connection string.}
}
\value{
A data frame with level ids, names and cutoffs.
}
\description{
Get level names and cut off values for a task
}
\examples{
getLevels('Sharing Standards 2017-2018 Year 5 Moderation Task A', 'mongodb://')
}
