% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getPages}
\alias{getPages}
\title{Get pages for a candidate}
\usage{
getPages(qrcode, task, connStr)
}
\arguments{
\item{qrcode}{QR Code}

\item{task}{Task id}

\item{connStr}{A connection string.}
}
\value{
A count of pages
}
\description{
Get pages for a candidate
}
\examples{
getPages('AF85HY','w2XzcMW3Rwb7pthe6', 'mongodb://')
}
