% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getPersons}
\alias{getPersons}
\title{Get persons from a task or tasks}
\usage{
getPersons(taskName, connStr)
}
\arguments{
\item{taskName}{The task name, tasks will be matched if their name includes the taskName string.}

\item{connStr}{A connection string.}
}
\value{
A data frame with persons
}
\description{
Get persons from a task or tasks
}
\examples{
getPersons('Sharing Standards 2017-2018 Year 3', 'mongodb://')
}
