% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{getPersonsFromSyllabus}
\alias{getPersonsFromSyllabus}
\title{Get candidates from a syllabus}
\usage{
getPersonsFromSyllabus(syllabus, connStr)
}
\arguments{
\item{syllabus}{The task syllabus}

\item{connStr}{A connection string.}
}
\value{
A data frame with persons
}
\description{
Get candidates from a syllabus
}
\examples{
getPersonsFromSyllabus('ee64c76a-5312-4e95-b315-ef5aca44539b', 'mongodb://')
}
