% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{getPersonsFromTask}
\alias{getPersonsFromTask}
\title{Get candidates from a task or tasks}
\usage{
getPersonsFromTask(task, mod = FALSE, connStr)
}
\arguments{
\item{task}{The task id}

\item{mod}{Is it a moderation task you want?}

\item{connStr}{A connection string.}
}
\value{
A data frame with persons
}
\description{
Get the candidates from a task. If the task is a moderation task, set mod=TRUE
}
\examples{
getPersonsFromTask('ee64c76a-5312-4e95-b315-ef5aca44539b',mod = TRUE, 'mongodb://')
}
