% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editor.R
\name{getPersonTexts}
\alias{getPersonTexts}
\title{Get texts from responses}
\usage{
getPersonTexts(ids, connStr)
}
\arguments{
\item{ids}{A vector of tasks}

\item{connStr}{A connection string.}
}
\value{
A data frame with comments
}
\description{
Get texts from responses
}
\examples{
getPersonTexts(ids,connStr)
}
