% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getProduct}
\alias{getProduct}
\title{Get product from a product id}
\usage{
getProduct(id, connStr)
}
\arguments{
\item{id}{The id of the product}

\item{connStr}{A connection string}
}
\description{
Get product from a product id
}
