% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{getProductByName}
\alias{getProductByName}
\title{Get product id from a product name}
\usage{
getProductByName(name, connStr)
}
\arguments{
\item{name}{The name of the product}

\item{connStr}{A connection string}
}
\description{
Get product id from a product name
}
