% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchases.R
\name{getPurchase}
\alias{getPurchase}
\title{Get purchase from a purchase id}
\usage{
getPurchase(id, connStr)
}
\arguments{
\item{id}{The id of the purchase}

\item{connStr}{A connection string}
}
\description{
Get purchase from a purchase id
}
