% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchases.R
\name{getPurchaseByProduct}
\alias{getPurchaseByProduct}
\title{Get purchase from a product id}
\usage{
getPurchaseByProduct(id, connStr)
}
\arguments{
\item{id}{The id of the product}

\item{connStr}{A connection string}
}
\description{
Get purchase from a product id
}
