% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{getSales}
\alias{getSales}
\title{Get sales data}
\usage{
getSales(product, connStr)
}
\arguments{
\item{product}{The product id}

\item{connStr}{A connection string.}
}
\value{
data frame product data
}
\description{
Get sales data
}
\examples{
\dontrun{
getSales('vqPwL2wWX8qc26nWs', 'mongodb://localhost')
}
}
