% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{getScaling}
\alias{getScaling}
\title{Get scaling factors for tasks}
\usage{
getScaling(task, connStr)
}
\arguments{
\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
data frame with _id, scaling.uScale and scaling.uiMean
}
\description{
Get scaling factors for tasks
}
\examples{
getScaling('FYB5kMkoh2v5sLsY7','mongodb://')
}
