% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{getSchoolByDfe}
\alias{getSchoolByDfe}
\title{Get school details from a dfe}
\usage{
getSchoolByDfe(dfe, connStr)
}
\arguments{
\item{dfe}{The dfe for the school}

\item{connStr}{A connection string.}
}
\value{
A data frame with school details
}
\description{
Get school details from a dfe
}
\examples{
getSchoolByDfe(999999999990, 'mongodb://')
}
