% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{getSchoolsByTrust}
\alias{getSchoolsByTrust}
\title{Get list of schools in a trust by trust and product name}
\usage{
getSchoolsByTrust(trustName, productName, connStr)
}
\arguments{
\item{trustName}{The trust name}

\item{productName}{The product name for the trust}

\item{connStr}{A connection string}
}
\value{
A data frame with school dfes
}
\description{
Get list of schools in a trust by trust and product name
}
\examples{
getSchoolsByTrust('Central South', 'mongodb://') 
}
