% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getScores}
\alias{getScores}
\title{Get scores from a task or tasks}
\usage{
getScores(taskName, connStr)
}
\arguments{
\item{taskName}{The task name, tasks will be matched if their name includes the taskName string.}

\item{connStr}{A connection string.}
}
\value{
A data frame with scores
}
\description{
Get scores from a task or tasks
}
\examples{
getScores('Sharing Standards 2017-2018 Year 3', 'mongodb://')
}
