% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus.R
\name{getSyllabusById}
\alias{getSyllabusById}
\title{Get syllabus object from a syllabus id}
\usage{
getSyllabusById(id, connStr)
}
\arguments{
\item{id}{The name for the syllabus}

\item{connStr}{A connection string.}
}
\value{
A data frame with syllabus object
}
\description{
Get syllabus object from a syllabus id
}
\examples{
getSyllabusById('abc', 'mongodb://') 
}
