% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus.R
\name{getSyllabusByName}
\alias{getSyllabusByName}
\title{Get syllabus object from a syllabus name}
\usage{
getSyllabusByName(name, connStr)
}
\arguments{
\item{name}{The name for the syllabus}

\item{connStr}{A connection string.}
}
\value{
A data frame with syllabus object
}
\description{
Get syllabus object from a syllabus name
}
\examples{
getSyllabusByName('SACS', 'mongodb://') will match all SACS syllabuses.
}
