% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getTask}
\alias{getTask}
\title{Get task id, name and number of decisions completed.}
\usage{
getTask(taskName, connStr)
}
\arguments{
\item{taskName}{Exact task name.}

\item{connStr}{A connection string.}
}
\value{
A data frame with task ids, task names and numbers of judgements.
}
\description{
Get task id, name and number of decisions completed.
}
\examples{
getTasks('Sharing Standards 2017-2018 Year 5', 'mongodb://') will match all Year 5 tasks.
}
