% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{getTaskById}
\alias{getTaskById}
\title{Get task object by id}
\usage{
getTaskById(id, connStr)
}
\arguments{
\item{id}{The task id.}

\item{connStr}{A connection string.}
}
\value{
A data frame with task details.
}
\description{
Get task object by id
}
\examples{
getTask('abc', 'mongodb://')
}
