% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{getTasksBySyllabus}
\alias{getTasksBySyllabus}
\title{Get task ids, names and number of decisions completed}
\usage{
getTasksBySyllabus(syllabus, connStr)
}
\arguments{
\item{syllabus}{The syllabus id for the tasks required}

\item{connStr}{A connection string.}
}
\value{
A data frame with task details.
}
\description{
Get task ids, names and number of decisions completed
}
\examples{
getTasks('7e606c3c-8e33-4c11-801f-6d168d79a869', 'mongodb://') will match all APW 2019 Year 3 tasks.
}
