% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{getTranscripts}
\alias{getTranscripts}
\title{Get audio transcripts by task ID}
\usage{
getTranscripts(ids, connStr)
}
\arguments{
\item{ids}{A character vector of task IDs.}

\item{connStr}{A connection string.}
}
\value{
A data frame with transcripts.
}
\description{
Get audio transcripts by task ID
}
\examples{
\dontrun{
getTranscripts('e4e50343-e73c-4542-a94e-230dce54096b', 'mongodb://localhost')
getTranscripts(c('e4e50343-e73c-4542-a94e-230dce54096b', 'abc123'), 'mongodb://localhost')
}
}
