% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\name{getTrusts}
\alias{getTrusts}
\title{Get trusts for a product}
\usage{
getTrusts(product, connStr)
}
\arguments{
\item{product}{The product id}

\item{connStr}{A connection string.}
}
\value{
A data frame with trust schools
}
\description{
Get trusts for a product
}
\examples{
getTrusts("abc", 'mongodb://')
}
