% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{getUser}
\alias{getUser}
\title{Get user details from an auth string}
\usage{
getUser(auth, connStr)
}
\arguments{
\item{auth}{The auth for the user}

\item{connStr}{A connection string.}
}
\value{
A data frame with user details
}
\description{
Get user details from an auth string
}
\examples{
getUser("auth0|607457cfe1af3d0072229628", 'mongodb://') 
}
