% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getUsers}
\alias{getUsers}
\title{Get user details}
\usage{
getUsers(connStr)
}
\arguments{
\item{connStr}{A connection string.}
}
\value{
A data frame with user details.
}
\description{
Get user details
}
\examples{
getUsers('mongodb://')
}
