% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{getWritingAgeSet}
\alias{getWritingAgeSet}
\title{Get writing age look up table}
\usage{
getWritingAgeSet(writingAgeSet, connStr)
}
\arguments{
\item{writingAgeSet}{The writing age set}

\item{connStr}{A connection string.}
}
\value{
The writing age set
}
\description{
Get the writingAgeSet look up table
}
\examples{
getWritingAgeSet('KS3-2021-2022', 'mongodb://')
}
