% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{humaniseDecisions}
\alias{humaniseDecisions}
\title{Recode AI decisions as human decisions}
\usage{
humaniseDecisions(task, connStr)
}
\arguments{
\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
updated count
}
\description{
Recode AI decisions as human decisions
}
