% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{removeAnchorProduct}
\alias{removeAnchorProduct}
\title{Remove anchor product for public anchor sets}
\usage{
removeAnchorProduct(task, connStr)
}
\arguments{
\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
List with modifiedCount, matchedCount, upsertedCount
}
\description{
Remove anchor product for public anchor sets
}
\examples{
removeAnchorProduct('FYB5kMkoh2v5sLsY7','mongodb://')
}
