% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{removeJudgingId}
\alias{removeJudgingId}
\title{Remove a judging pair with a judging id from getJudging}
\usage{
removeJudgingId(judgingid, connStr)
}
\arguments{
\item{judgingid}{The name of the pair to remove}

\item{connStr}{A connection string.}
}
\description{
Remove a judging pair with a judging id from getJudging
}
\examples{
removeJudgingId('gpftTfdPKeBEiz7pd', 'mongodb://')
}
