% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{resetAiFeedback}
\alias{resetAiFeedback}
\title{Reset AI feedback for a task}
\usage{
resetAiFeedback(task, connStr)
}
\arguments{
\item{task}{The task id.}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Removes aiReports from all candidates with the given task and removes
schoolSummary and teacherGroupSummaries from the task itself.
Also deletes all quiz questions associated with the task.
}
\examples{
resetAiFeedback('da3a061f-7813-41fc-b7df-f7ec4a41e3c1', 'mongodb://')
}
