% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{resetDecisions}
\alias{resetDecisions}
\title{Re-include all decisions from a task}
\usage{
resetDecisions(task, connStr)
}
\arguments{
\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
updated count
}
\description{
Re-include all decisions from a task
}
