% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{resetModJudges}
\alias{resetModJudges}
\title{Re-include judges in a moderation task}
\usage{
resetModJudges(task, connStr)
}
\arguments{
\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
Records affected
}
\description{
Re-include judges in a moderation task
}
\examples{
resetModJudges('o544AXuwEBADM9Lrk', 'mongodb://')
}
