% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisions.R
\name{robotiseDecisions}
\alias{robotiseDecisions}
\title{Recode AI decisions as robot decisions}
\usage{
robotiseDecisions(task, connStr)
}
\arguments{
\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
updated count
}
\description{
Recode AI decisions as robot decisions
}
