% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{setAsAnchor}
\alias{setAsAnchor}
\title{Update candidate anchor values}
\usage{
setAsAnchor(firstName, lastName, modTask, theta, connStr)
}
\arguments{
\item{firstName}{the name of the anchor}

\item{lastName}{a label for the anchor}

\item{modTask}{the id of the moderation task}

\item{theta}{the anchor value of the script}

\item{connStr}{A connection string.}
}
\value{
the updated records
}
\description{
Update candidate anchoring properties in a mod task
}
