% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{setLevels}
\alias{setLevels}
\title{Set levels data}
\usage{
setLevels(task, boundaries, connStr)
}
\arguments{
\item{task}{The task id}

\item{boundaries}{A dataframe with the columns: task, name, value, include, cutOff}

\item{connStr}{A connection string.}
}
\value{
The levels that were updated
}
\description{
Set levels data
}
\examples{
setLevels('74537cb5-d97e-4d48-b89a-97a81b91cd34','mongodb://')
}
