% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{setModerationCode}
\alias{setModerationCode}
\title{Set a moderation code for a task}
\usage{
setModerationCode(task, modCode, connStr)
}
\arguments{
\item{task}{The task id.}

\item{modCode}{The moderation code.}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Set a moderation code for a task
}
\examples{
setModerationCode('FJRRtRsNxm7TG93o9','sZBLjuzPgiJwonDTw', 'mongodb://')
}
