% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus.R
\name{setSyllabusReadyForJudging}
\alias{setSyllabusReadyForJudging}
\title{Set ready for judging on a syllabus}
\usage{
setSyllabusReadyForJudging(syllabus, ready, connStr)
}
\arguments{
\item{syllabus}{The syllabus id}

\item{ready}{Set ready or not. Boolean.}

\item{connStr}{A connection string.}
}
\value{
List with modifiedCount, matchedCount, upsertedCount
}
\description{
Set ready for judging on a syllabus
}
\examples{
setSyllabusReadyForJudging('ad888bb7-e47a-4e6b-b827-1498c752a389',TRUE, 'mongodb://')
}
