% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{unsetModAnchors}
\alias{unsetModAnchors}
\title{Unset anchors in a mod task}
\usage{
unsetModAnchors(modTask, connStr)
}
\arguments{
\item{modTask}{the name of the mod task}

\item{connStr}{A connection string.}
}
\value{
the updated records
}
\description{
Set anchor = false for all candidates in a mod task
}
