% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{updateAnchor}
\alias{updateAnchor}
\title{Update anchor value for a person id}
\usage{
updateAnchor(id, anchorValue, connStr)
}
\arguments{
\item{id}{The person id}

\item{anchorValue}{The anchor value}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Update anchor value for a person id
}
\examples{
updateAnchor('o544AXuwEBADM9Lrk',0.27, 'mongodb://')
}
