% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{updateAnchorProduct}
\alias{updateAnchorProduct}
\title{Set anchor product for public anchor sets}
\usage{
updateAnchorProduct(task, productName, connStr)
}
\arguments{
\item{task}{The task id}

\item{productName}{The product name}

\item{connStr}{A connection string.}
}
\value{
List with modifiedCount, matchedCount, upsertedCount
}
\description{
Set anchor product for public anchor sets
}
\examples{
updateAnchorProduct('FYB5kMkoh2v5sLsY7','APW', 'mongodb://')
}
