% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{updateDefaultJudgements}
\alias{updateDefaultJudgements}
\title{Update judge quota}
\usage{
updateDefaultJudgements(task, quota, connStr)
}
\arguments{
\item{task}{The task id}

\item{quota}{The default quota}

\item{connStr}{A connection string.}
}
\value{
List with modifiedCount, matchedCount, upsertedCount
}
\description{
Update judge quota
}
\examples{
updateDefaultJudgements('FYB5kMkoh2v5sLsY7',55, 'mongodb://')
}
