% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{updateExport}
\alias{updateExport}
\title{Set script as export into anchor set}
\usage{
updateExport(name, task, connStr)
}
\arguments{
\item{name}{The script name}

\item{task}{The task id}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Set script as export into anchor set
}
\examples{
updateExport('XcuTh8dh9ontyKj86','7CHCVQ', 'mongodb://')
}
