% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{updateFromSyllabus}
\alias{updateFromSyllabus}
\title{Set fromSyllabus for a task}
\usage{
updateFromSyllabus(task, syllabus, connStr)
}
\arguments{
\item{task}{The task id}

\item{syllabus}{The syllabus id}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Set fromSyllabus for a task
}
\examples{
updateFromSyllabus('FYB5kMkoh2v5sLsY7','FYB5kMkoh2v5sLsY8', 'mongodb://')
}
