% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judges.R
\name{updateJudgeQuota}
\alias{updateJudgeQuota}
\title{Update an individual judge quota}
\usage{
updateJudgeQuota(judge, quota, connStr)
}
\arguments{
\item{judge}{The judge id}

\item{quota}{The new quota}

\item{connStr}{A connection string.}
}
\value{
List with modifiedCount, matchedCount, upsertedCount
}
\description{
Update an individual judge quota
}
\examples{
updateJudgeQuota('ad888bb7-e47a-4e6b-b827-1498c752a389',55, 'mongodb://')
}
