% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{updateMark}
\alias{updateMark}
\title{Update the mark property of a candidate}
\usage{
updateMark(candidateId, mark, connStr)
}
\arguments{
\item{candidateId}{The id of the candidate}

\item{mark}{The mark to set}

\item{connStr}{A connection string.}
}
\value{
The updated record
}
\description{
Update the mark property of a candidate
}
