% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{updateScaling}
\alias{updateScaling}
\title{Set scaling factors for tasks}
\usage{
updateScaling(task, uScale, uiMean, connStr)
}
\arguments{
\item{task}{The task id}

\item{uScale}{Scaling factor}

\item{uiMean}{Mean of scale}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Set scaling factors for tasks
}
\examples{
updateScaling('FYB5kMkoh2v5sLsY7',6, 55, 'mongodb://')
}
