% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{updateShiftConstant}
\alias{updateShiftConstant}
\title{Update anchor constant}
\usage{
updateShiftConstant(modTask, shiftConstant, connStr)
}
\arguments{
\item{modTask}{The mod task id}

\item{shiftConstant}{The shift constant}

\item{connStr}{A connection string.}
}
\value{
List with modifiedCount, matchedCount, upsertedCount
}
\description{
Update anchor constant
}
\examples{
updateShiftConstant('FYB5kMkoh2v5sLsY7',3, 'mongodb://')
}
