% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoices.R
\name{createInvoice}
\alias{createInvoice}
\title{Create a new invoice}
\usage{
createInvoice(dfe, product, purchase_id, no, connStr)
}
\arguments{
\item{dfe}{dfe number}

\item{product}{product id}

\item{no}{invoice number}

\item{connStr}{A connection string.}

\item{purchase}{purchase id}
}
\value{
A string with confirmation
}
\description{
Create a new invoice
}
\examples{
createInvoice(dfe, product, purchase, no, 'mongodb://')
}
