% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getDecisions}
\alias{getDecisions}
\title{Get decisions from a task or tasks}
\usage{
getDecisions(taskId, connStr)
}
\arguments{
\item{taskId}{The task ID.}

\item{connStr}{A connection string.}
}
\value{
A data frame with decisions.
}
\description{
Get decisions from a task or tasks
}
\examples{
getDecisions('67059790-569b-440f-a858-bddabf313e07', 'mongodb://')
}
