% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getJudges}
\alias{getJudges}
\title{Get judge details}
\usage{
getJudges(taskId, modTask = FALSE, connStr)
}
\arguments{
\item{taskId}{Task Id}

\item{modTask}{Is the task id a moderation task, TRUE or FALSE}

\item{connStr}{A connection string.}
}
\value{
A data frame with judge details.
}
\description{
Get judge details
}
\examples{
getJudges(localTask,modTask=FALSE,connStr)
getJudges(modTask,modTask=TRUE,connStr)
}
