% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getJudges}
\alias{getJudges}
\title{Get judge details}
\usage{
getJudges(taskId, connStr)
}
\arguments{
\item{taskId}{Task Id}

\item{connStr}{A connection string.}
}
\value{
A data frame with user details.
}
\description{
Get judge details
}
\examples{
getJudges('ocjZYZAYLuZJyAv4h', 'mongodb://')
}
