% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getPersons}
\alias{getPersons}
\title{Get persons from a task or tasks}
\usage{
getPersons(syllabus, connStr)
}
\arguments{
\item{syllabus}{The task syllabus}

\item{connStr}{A connection string.}
}
\value{
A data frame with persons
}
\description{
Get persons from a task or tasks
}
\examples{
getPersons('ee64c76a-5312-4e95-b315-ef5aca44539b', 'mongodb://')
}
