% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\name{getSales}
\alias{getSales}
\title{Get sales data}
\usage{
getSales(product, connStr)
}
\arguments{
\item{connStr}{A connection string.}

\item{project}{The project id}
}
\value{
data frame product data
}
\description{
Get sales data
}
\examples{
getSales('vqPwL2wWX8qc26nWs','mongodb://')
}
