% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getSales}
\alias{getSales}
\title{Get sales data. Updated for nmm-vegas-db.}
\usage{
getSales(product, connStr)
}
\arguments{
\item{connStr}{A connection string.}

\item{project}{The project id}
}
\value{
data frame product data
}
\description{
Get sales data. Updated for nmm-vegas-db.
}
\examples{
getSales('vqPwL2wWX8qc26nWs','mongodb://')
}
