% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getSchoolByDfe}
\alias{getSchoolByDfe}
\title{Get school details from a dfe}
\usage{
getSchoolByDfe(dfe, connStr)
}
\arguments{
\item{dfe}{The dfe for the school}

\item{connStr}{A connection string.}
}
\value{
A data frame with school names and urns
}
\description{
Get school details from a dfe
}
\examples{
getSchoolByName(999999999990, 'mongodb://') will find the school with the dfe 999999999990.
}
