% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getSyllabusByName}
\alias{getSyllabusByName}
\title{Get syllabus ids from a syllabus name}
\usage{
getSyllabusByName(name, connStr)
}
\arguments{
\item{name}{The name for the syllabus}

\item{connStr}{A connection string.}
}
\value{
A data frame with syllabus ids
}
\description{
Get syllabus ids from a syllabus name
}
\examples{
getSyllabusByName('SACS', 'mongodb://') will match all SACS syllabuses.
}
