% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{getTasks}
\alias{getTasks}
\title{Get task ids, names and number of decisions completed. Updated for nmm-vegas-db.}
\usage{
getTasks(taskName, connStr)
}
\arguments{
\item{taskName}{Tasks will be matched if their name includes the taskName string.}

\item{connStr}{A connection string.}
}
\value{
A data frame with task details.
}
\description{
Get task ids, names and number of decisions completed. Updated for nmm-vegas-db.
}
\examples{
getTasks('Sharing Standards 2017-2018 Year 5', 'mongodb://') will match all Year 5 tasks.
}
