% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\name{getTranscripts}
\alias{getTranscripts}
\title{Get audio transcripts for a task}
\usage{
getTranscripts(taskId, connStr)
}
\arguments{
\item{taskId}{The task ID.}

\item{connStr}{A connection string.}
}
\value{
A data frame with transcripts.
}
\description{
Get audio transcripts for a task
}
\examples{
\dontrun{
getTranscripts('e4e50343-e73c-4542-a94e-230dce54096b', 'mongodb://localhost')
}
}
