% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{removeJudgingId}
\alias{removeJudgingId}
\title{Get pages for a candidate}
\usage{
removeJudgingId(judgingid, connStr)
}
\arguments{
\item{judgingid}{The name of the pair to remove}

\item{connStr}{A connection string.}
}
\description{
Get pages for a candidate
}
\examples{
removeJudgingId('gpftTfdPKeBEiz7pd', 'mongodb://')
}
