% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persons.R
\name{updateAnchor}
\alias{updateAnchor}
\title{Update candidate anchor values}
\usage{
updateAnchor(id, theta, connStr)
}
\arguments{
\item{id}{the id of the anchor}

\item{theta}{the anchor value of the script}

\item{connStr}{A connection string.}
}
\value{
the updated records
}
\description{
Update anchor scores in a mod task
}
