% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite-connection.R
\name{updateScaling}
\alias{updateScaling}
\title{Set anchor scaling factors for tasks}
\usage{
updateScaling(task, uScale, uiMean, anchorScale = TRUE, connStr)
}
\arguments{
\item{task}{The task id}

\item{uScale}{Anchor scaling factor}

\item{uiMean}{Mean of anchor scale}

\item{anchorScale}{Anchor scaling or normal scaling?}

\item{connStr}{A connection string.}
}
\value{
Nothing
}
\description{
Set anchor scaling factors for tasks
}
\examples{
updateScaling('FYB5kMkoh2v5sLsY7',6, 55, TRUE, 'mongodb://')
}
